Created a new level to work on and the transition between levels
Put a trigger volume in the level, then go into the level blueprint, use OnActorBeginOverlap and CastToTriggerVolume to connect to Open Level (By Name). Make sure you put in the exact name of the level you want to open.
If need be, we can roll back commits if we screw something up.